Refactored the code according to the codeRabbit review
